// cmp_mini_frontend/src/servers/models/AddressVO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type AddressVO = {
    cityName?: string;
    countyName?: string;
    detail?: string;
    id?: number;
    isDefault?: number;
    name?: string;
    phone?: string;
    provinceName?: string;
    townName?: string;
};